# apps/api/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: Tanban Restaurant SaaS API
  version: 0.1.0
  description: 所有业务时间均按北京时间（Asia/Shanghai，UTC+8）输入、计算和展示；响应中的本地日期时间统一为 `YYYY-MM-DD HH:mm:ss`。需表达绝对时刻的请求可使用带 `+08:00` 的 RFC3339。
servers:
  - url: https://tbapi.666qwe.cn/api/v1
paths:
  /auth/login:
    post:
      summary: 平台或商户账号登录
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username, password, portal]
              properties:
                username: {type: string}
                password: {type: string, format: password}
                portal: {type: string, enum: [platform, merchant]}
      responses:
        '200': {description: JWT and current user}
  /auth/me:
    get:
      summary: 当前账号
      security: [{bearerAuth: []}]
      responses:
        '200': {description: Current user}
...
  /merchant/marketing/apps:
    get:
      summary: 已开放营销应用及活动状态汇总
      security: [{bearerAuth: []}]
      responses: {'200': {description: Marketing application registry}}
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...